Clean Exit

+++ src/io.h
@@ -11,7 +11,7 @@ void init_io(int width, int height);
 void deinit_io();
 
 int draw_io(int x, int y, int n, uint8_t * mem);
-uint16_t read_keys_io();
+uint32_t read_keys_io();
 void clearscreen_io();
 void refresh_io();
